// packages/confluence/src/confluence-client/models/SpacePermission.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OperationDescription } from './OperationDescription.js';
import type { Subject } from './Subject.js';
export type SpacePermission = {
    operation?: OperationDescription;
    subject?: Subject;
    spaceKey?: string;
    spaceId?: number;
};