// packages/confluence/src/confluence-client/models/SpacePermissionsForSubject.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { OperationDescription } from './OperationDescription.js';
export type SpacePermissionsForSubject = {
    userKey?: string;
    groupName?: string;
    operations?: Array<OperationDescription>;
};